+++ mycli/main.py
@@ -24,6 +24,7 @@ from cli_helpers.tabular_output import preprocessors
 from cli_helpers.utils import strip_ansi
 import click
 import sqlparse
+import sqlglot
 from mycli.packages.parseutils import is_dropping_database, is_destructive
 from prompt_toolkit.completion import DynamicCompleter
 from prompt_toolkit.enums import DEFAULT_BUFFER, EditingMode
@@ -123,6 +124,7 @@ class MyCli(object):
         self.logfile = logfile
         self.defaults_suffix = defaults_suffix
         self.login_path = login_path
+        self.toolbar_error_message = None
 
         # self.cnf_files is a class variable that stores the list of mysql
         # config files to read in at launch.
@@ -581,6 +583,34 @@ class MyCli(object):
                 raise RuntimeError(message)
             return True
         return False
+
+    def handle_prettify_binding(self, text):
+        try:
+            statements = sqlglot.parse(text, read='mysql')
+        except Exception as e:
+            statements = []
+        if len(statements) == 1:
+            pretty_text = statements[0].sql(pretty=True, pad=4, dialect='mysql')
+        else:
+            pretty_text = ''
+            self.toolbar_error_message = 'Prettify failed to parse statement'
+        if len(pretty_text) > 0:
+            pretty_text = pretty_text + ';'
+        return pretty_text
+
+    def handle_unprettify_binding(self, text):
+        try:
+            statements = sqlglot.parse(text, read='mysql')
+        except Exception as e:
+            statements = []
+        if len(statements) == 1:
+            unpretty_text = statements[0].sql(pretty=False, dialect='mysql')
+        else:
+            unpretty_text = ''
+            self.toolbar_error_message = 'Unprettify failed to parse statement'
+        if len(unpretty_text) > 0:
+            unpretty_text = unpretty_text + ';'
+        return unpretty_text
 
     def run_cli(self):
         iterations = 0